+++ Staff/Staff.Domain/BaseModels/EntityWithId.cs
@@ -1,0 +1,12 @@
+﻿namespace Staff.Domain.BaseModels;
+
+/// <summary>
+/// Базовый класс для сущностей с идентификатором.
+/// </summary>
+public abstract class EntityWithId
+{
+    /// <summary>
+    /// Уникальный идентификатор сущности.
+    /// </summary>
+    public int Id { get; set; }
+}
