+++ Staff/Staff.Domain/Enums/Gender.cs
@@ -1,4 +1,4 @@
-﻿namespace Staff.Domain;
+﻿namespace Staff.Domain.Enums;
 
 /// <summary>
 /// Перечисление, представляющее пол сотрудника.
