+++ Staff/Staff.Domain/Models/EmploymentArchiveRecord.cs
@@ -1,4 +1,6 @@
-﻿namespace Staff.Domain;
+﻿using Staff.Domain.Enums;
+
+namespace Staff.Domain.Models;
 
 /// <summary>
 /// Представляет запись в архиве трудовой деятельности сотрудника.
