+++ Staff/Staff.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 17.11.35327.3
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Staff.Domain", "Staff.Domain\Staff.Domain.csproj", "{02B1C7D0-B080-479B-825C-2836992EA062}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Staff.Tests", "Staff.Tests\Staff.Tests.csproj", "{E41F8171-E915-451C-B13F-5F3AB2EE6AF7}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{02B1C7D0-B080-479B-825C-2836992EA062}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{02B1C7D0-B080-479B-825C-2836992EA062}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{02B1C7D0-B080-479B-825C-2836992EA062}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E41F8171-E915-451C-B13F-5F3AB2EE6AF7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E41F8171-E915-451C-B13F-5F3AB2EE6AF7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E41F8171-E915-451C-B13F-5F3AB2EE6AF7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E41F8171-E915-451C-B13F-5F3AB2EE6AF7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
